+++ apps/cow-fi/components/Layout/index.tsx
@@ -2,13 +2,13 @@
 
 import { PropsWithChildren, ReactNode } from 'react'
 
-import { Footer, GlobalCoWDAOStyles, Media, MenuBar } from '@cowprotocol/ui'
-import { baseTheme } from '@cowprotocol/ui'
+import { useFeatureFlags } from '@cowprotocol/common-hooks'
+import { Footer, GlobalCoWDAOStyles, Media, MenuBar, baseTheme } from '@cowprotocol/ui'
 
 import Link from 'next/link'
 import styled, { createGlobalStyle, css, ThemeProvider } from 'styled-components/macro'
 
-import { NAV_ADDITIONAL_BUTTONS, NAV_ITEMS, PAGE_MAX_WIDTH, PRODUCT_VARIANT } from './const'
+import { getNavItems, NAV_ADDITIONAL_BUTTONS, PAGE_MAX_WIDTH, PRODUCT_VARIANT } from './const'
 
 import { useSetupPage } from '../../hooks/useSetupPage'
 import { CowSaucerScene } from '../CowSaucerScene'
@@ -44,6 +44,7 @@ interface LayoutProps {
 
 export function Layout({ children, bgColor, host, showCowSaucer, contentMinHeight }: Readonly<LayoutProps>): ReactNode {
   useSetupPage()
+  const { isSolversEnabled } = useFeatureFlags()
 
   const GlobalStyles = GlobalCoWDAOStyles()
   const LocalStyles = createGlobalStyle(
@@ -61,7 +62,7 @@ export function Layout({ children, bgColor, host, showCowSaucer, contentMinHeigh
       {/* Override global light theme to force dark mode for MenuBar only */}
       <ThemeProvider theme={darkTheme}>
         <MenuBar
-          navItems={NAV_ITEMS}
+          navItems={getNavItems(!!isSolversEnabled)}
           productVariant={PRODUCT_VARIANT}
           additionalNavButtons={NAV_ADDITIONAL_BUTTONS}
           padding="10px 60px"