+++ apps/cowswap-frontend/src/modules/twap/services/cancelTwapOrderTxs.ts
@@ -1,19 +1,7 @@
-import { USDC_LENS, WRAPPED_NATIVE_CURRENCIES } from '@cowprotocol/common-const'
 import { isZkSyncChain, SupportedChainId } from '@cowprotocol/cow-sdk'
 import { ComposableCoW, GPv2Settlement } from '@cowprotocol/cowswap-abis'
-import { CurrencyAmount } from '@cowprotocol/currency'
-import { ContractsOrder } from '@cowprotocol/sdk-contracts-ts'
 import { BigNumber } from '@ethersproject/bignumber'
 import type { MetaTransactionData } from '@safe-global/types-kit'
-
-import { toKeccak256 } from 'common/utils/toKeccak256'
-
-import { computeOrderUid } from '../../../utils/orderUtils/computeOrderUid'
-import { TWAPOrder, TwapOrderItem } from '../types'
-import { buildTwapOrderParamsStruct } from '../utils/buildTwapOrderParamsStruct'
-import { createPartOrderFromParent } from '../utils/buildTwapParts'
-import { getConditionalOrderId } from '../utils/getConditionalOrderId'
-import { twapOrderToStruct } from '../utils/twapOrderToStruct'
 
 export interface CancelTwapOrderContext {
   composableCowContract: ComposableCoW
@@ -47,9 +35,11 @@ export function cancelTwapOrderTxs(context: CancelTwapOrderContext): MetaTransac
 // TODO: we might need a custom method for estimating gas on Linea
 export async function estimateCancelTwapOrderTxs(context: CancelTwapOrderContext): Promise<BigNumber> {
   if (isZkSyncChain(context.chainId)) {
-    // Lens is a zkSync chain, so we need to use a different estimation method
+    throw new Error('estimateCancelTwapOrderTxs: Please, re-enable zkSync chain estimation.')
+
+    // We need to use a different estimation method for zkSync chains (like we did for Lens).
     // See the function estimateZkSyncCancelTwapOrderTxs for details
-    return estimateZkSyncCancelTwapOrderTxs(context)
+    // return estimateZkSyncCancelTwapOrderTxs(context)
   }
 
   const { composableCowContract, settlementContract, orderId, partOrderId } = context
@@ -71,6 +61,7 @@ export async function estimateCancelTwapOrderTxs(context: CancelTwapOrderContext
  *
  * Thus, this function estimates the gas cost by simulating the transaction as if it were sent by a fake EOA address.
  */
+/*
 async function estimateZkSyncCancelTwapOrderTxs(context: CancelTwapOrderContext): Promise<BigNumber> {
   const { composableCowContract, settlementContract, partOrderId: hasPartOrder, chainId } = context
 
@@ -100,8 +91,13 @@ const FAKE_OWNER = '0x330d9F4906EDA1f73f668660d1946bea71f48827'
 
 const START_TIME = Math.floor(Date.now() / 1000)
 
+const zkSyncChain: ChainInfo =
+  ALL_SUPPORTED_CHAINS.find(({ id }) => isZkSyncChain(id)) || ALL_SUPPORTED_CHAINS_MAP[SupportedChainId.MAINNET]
+
+const zkSyncChainId = zkSyncChain.id as SupportedChainId
+
 const FAKE_TWAP_ORDER: TWAPOrder = {
-  sellAmount: CurrencyAmount.fromRawAmount(WRAPPED_NATIVE_CURRENCIES[SupportedChainId.LENS], 100_000_000_000),
+  sellAmount: CurrencyAmount.fromRawAmount(WRAPPED_NATIVE_CURRENCIES[zkSyncChainId], 100_000_000_000),
   buyAmount: CurrencyAmount.fromRawAmount(USDC_LENS, 200_000_000_000),
   receiver: FAKE_OWNER,
   numOfParts: 2,
@@ -144,3 +140,4 @@ async function getFakeTwapPartOrderId(chainId: SupportedChainId): Promise<string
   FAKE_PART_ORDER_IDS_CACHE[chainId] = await computeOrderUid(chainId, FAKE_OWNER, part as ContractsOrder)
   return FAKE_PART_ORDER_IDS_CACHE[chainId]
 }
+*/